refactor: return config from ping
