+++ packages/core-p2p/src/interfaces.ts
@@ -3,18 +3,3 @@ export interface IPeerData {
     port: number;
     version: string;
 }
-
-export interface IResponseConfig {
-    version: string;
-    network: {
-        version: string;
-        name: string;
-        nethash: string;
-        explorer: string;
-        token: {
-            name: string;
-            symbol: string;
-        };
-    };
-    plugins: { [key: string]: { enabled: boolean; port: number } };
-}
